refactor: removed legacy tests
